Extract layergroupIdHeader middleware
