+++ lib/cartodb/controllers/map/anonymous.js
@@ -17,6 +17,7 @@ const surrogateKeyHeader = require('../../middleware/surrogate-key-header');
 const lastModifiedHeader = require('../../middleware/last-modified-header');
 const lastUpdatedTimeLayergroup = require('./middlewares/last-updated-time-layergroup');
 const layerStats = require('./middlewares/layer-stats');
+const layergroupIdHeader = require('./middlewares/layergroup-id-header');
 const sendResponse = require('../../middleware/send-response');
 const CreateLayergroupMapConfigProvider = require('../../models/mapconfig/provider/create-layergroup-provider');
 const LayergroupMetadata = require('../../utils/layergroup-metadata');
@@ -103,7 +104,7 @@ AnonymousMapController.prototype.composeCreateMapMiddleware = function () {
         lastModifiedHeader({ now: true }),
         lastUpdatedTimeLayergroup(),
         layerStats(this.pgConnection, this.statsBackend),
-        setLayergroupIdHeader(this.templateMaps ,useTemplateHash),
+        layergroupIdHeader(this.templateMaps ,useTemplateHash),
         setDataviewsAndWidgetsUrlsToLayergroupMetadata(this.layergroupMetadata),
         setAnalysesMetadataToLayergroup(this.layergroupMetadata, includeQuery),
         setTurboCartoMetadataToLayergroup(this.layergroupMetadata),
@@ -215,22 +216,6 @@ function createLayergroup (mapBackend, userLimitsApi, pgConnection, affectedTabl
 
             next();
         });
-    };
-}
-
-function setLayergroupIdHeader (templateMaps, useTemplateHash) {
-    return function setLayergroupIdHeaderMiddleware (req, res, next) {
-        const { user, template } = res.locals;
-        const layergroup = res.body;
-
-        if (useTemplateHash) {
-            var templateHash = templateMaps.fingerPrint(template).substring(0, 8);
-            layergroup.layergroupid = `${user}@${templateHash}@${layergroup.layergroupid}`;
-        }
-
-        res.set('X-Layergroup-Id', layergroup.layergroupid);
-
-        next();
     };
 }
