+++ lib/cartodb/controllers/map/named.js
@@ -14,6 +14,7 @@ const surrogateKeyHeader = require('../../middleware/surrogate-key-header');
 const lastModifiedHeader = require('../../middleware/last-modified-header');
 const lastUpdatedTimeLayergroup = require('./middlewares/last-updated-time-layergroup');
 const layerStats = require('./middlewares/layer-stats');
+const layergroupIdHeader = require('./middlewares/layergroup-id-header');
 const sendResponse = require('../../middleware/send-response');
 const NamedMapMapConfigProvider = require('../../models/mapconfig/provider/named-map-provider');
 const CreateLayergroupMapConfigProvider = require('../../models/mapconfig/provider/create-layergroup-provider');
@@ -114,7 +115,7 @@ NamedMapController.prototype.composeInstantiateTemplateMiddleware = function () 
         lastModifiedHeader({ now: true }),
         lastUpdatedTimeLayergroup(),
         layerStats(this.pgConnection, this.statsBackend),
-        setLayergroupIdHeader(this.templateMaps ,useTemplateHash),
+        layergroupIdHeader(this.templateMaps ,useTemplateHash),
         setDataviewsAndWidgetsUrlsToLayergroupMetadata(this.layergroupMetadata),
         setAnalysesMetadataToLayergroup(this.layergroupMetadata, includeQuery),
         setTurboCartoMetadataToLayergroup(this.layergroupMetadata),
@@ -222,22 +223,6 @@ function instantiateLayergroup (mapBackend, userLimitsApi, pgConnection, affecte
 
             next();
         });
-    };
-}
-
-function setLayergroupIdHeader (templateMaps, useTemplateHash) {
-    return function setLayergroupIdHeaderMiddleware (req, res, next) {
-        const { user, template } = res.locals;
-        const layergroup = res.body;
-
-        if (useTemplateHash) {
-            var templateHash = templateMaps.fingerPrint(template).substring(0, 8);
-            layergroup.layergroupid = `${user}@${templateHash}@${layergroup.layergroupid}`;
-        }
-
-        res.set('X-Layergroup-Id', layergroup.layergroupid);
-
-        next();
     };
 }
